+++ src/App.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import Calendar from './components/Calendar/Calendar';
+import './App.css';
 
 function App() {
   return <Calendar />;
